99% homework

+++ app/src/main/java/com/example/doancn/Models/HomeWorkX.kt
@@ -5,7 +5,7 @@ import java.io.Serializable
 data class HomeWorkX(
     val dateCreated: String,
     val deadline: String,
-    val fileId: Int,
+    val fileId: Long,
     val name: String,
     val sizeInByte: Int
 ) : Serializable
